Issue #2 move current caching tools to dedicated `caching` module
