+++ src/openeo_aggregator/utils.py
@@ -1,90 +1,18 @@
 import datetime
 import functools
 import logging
-import types
-
-import shapely.geometry
 import time
 from typing import Callable, Iterable, Iterator, List, NamedTuple, Set
 
-from openeo.util import TimingLogger, rfc3339
+import shapely.geometry
+
+from openeo.util import rfc3339
 
 # Generic "sentinel object" for unset values (where `None` is valid value)
 # https://python-patterns.guide/python/sentinel-object/)
 _UNSET = object()
 
-
-class CacheMissException(KeyError):
-    pass
-
-
 _log = logging.getLogger(__name__)
-
-
-class TtlCache:
-    """
-    Simple (in-memory) dict-style cache with expiry
-    """
-
-    def __init__(self, default_ttl: int = 60, clock: Callable[[], float] = time.time, name: str = None):
-        self._cache = {}
-        self.default_ttl = default_ttl
-        self._clock = clock  # TODO: centralized helper for this test pattern
-        self.name = name or "TtlCache"
-
-    def set(self, key, value, ttl=None):
-        """Add item to cache"""
-        self._cache[key] = (value, self._clock() + (ttl or self.default_ttl))
-
-    def __setitem__(self, key, value):
-        """Add item to cache"""
-        self.set(key, value)
-
-    def contains(self, key) -> bool:
-        """Check whether cache contains item under given key"""
-        if key not in self._cache:
-            return False
-        value, expiration = self._cache[key]
-        if self._clock() <= expiration:
-            return True
-        else:
-            del self._cache[key]
-            return False
-
-    def __contains__(self, key):
-        """Check whether cache contains item under given key"""
-        return self.contains(key)
-
-    def get(self, key, default=None):
-        """Get item from cache and if not available: return default value."""
-        return self._cache[key][0] if self.contains(key) else default
-
-    def __getitem__(self, key):
-        """Get item from cache and raise `CacheMissException` if not available."""
-        if self.contains(key):
-            return self._cache[key][0]
-        raise CacheMissException(key)
-
-    def get_or_call(self, key, callback, ttl=None, log_on_miss=False):
-        """
-        Helper to compactly implement the "get from cache or calculate otherwise" pattern
-        """
-        if self.contains(key):
-            res = self[key]
-        else:
-            if log_on_miss:
-                with TimingLogger(
-                        title=f"Cache miss {self.name!r} key {key!r}, calling {callback.__qualname__!r}",
-                        logger=_log.debug
-                ):
-                    res = callback()
-            else:
-                res = callback()
-            self.set(key, res, ttl=ttl)
-        return res
-
-    def flush_all(self):
-        self._cache = {}
 
 
 class MultiDictGetter:
